#38 Extracted Instances to another package.
