+++ src/test/java/com/jcabi/dynamodb/maven/plugin/InstancesTest.java
@@ -41,6 +41,7 @@ import com.amazonaws.services.dynamodbv2.model.KeyType;
 import com.amazonaws.services.dynamodbv2.model.ProvisionedThroughput;
 import com.amazonaws.services.dynamodbv2.model.PutItemRequest;
 import com.amazonaws.services.dynamodbv2.model.ScalarAttributeType;
+import com.jcabi.dynamodb.core.Instances;
 import java.io.File;
 import java.net.ServerSocket;
 import java.util.Collections;
